Allow stuff to be removed
This commit entails the removal of content from the store.  Please note
that there is currently no cleaning up after the elements in the store.
Hence removing a catalog may leave some clutter around.

Why? You ask?  Custom business logic should decide that the related
datasets etc need to be removed.  Adding such logic is a long and future
goal of mu-star-resources, but it is currently not implemented.